+++ cpp/src/compute/Backend.cc
@@ -15,16 +15,25 @@
  * limitations under the License.
  */
 
-#include "result_iterator.h"
-#include "exec_backend.h"
+#include "Backend.h"
 
 namespace gluten {
 
-std::shared_ptr<Metrics> ResultIterator::GetMetrics() {
-  if (backend_) {
-    return backend_->GetMetrics(raw_iter_, exportNanos_);
+static std::function<std::shared_ptr<Backend>()> backend_factory;
+
+void SetBackendFactory(std::function<std::shared_ptr<Backend>()> factory) {
+#ifdef GLUTEN_PRINT_DEBUG
+  std::cout << "Set backend factory." << std::endl;
+#endif
+  backend_factory = std::move(factory);
+}
+
+std::shared_ptr<Backend> CreateBackend() {
+  if (backend_factory == nullptr) {
+    throw std::runtime_error(
+        "Execution backend not set. This may due to the backend library not loaded, or SetBackendFactory() is not called in nativeInitNative() JNI call.");
   }
-  return nullptr;
+  return backend_factory();
 }
 
 } // namespace gluten
