+++ cpp/src/compute/Backend.h
@@ -22,14 +22,14 @@
 #include <arrow/record_batch.h>
 #include <arrow/util/iterator.h>
 
-#include "compute/protobuf_utils.h"
-#include "compute/result_iterator.h"
-#include "memory/arrow_memory_pool.h"
-#include "memory/columnar_batch.h"
-#include "operators/c2r/arrow_columnar_to_row_converter.h"
+#include "compute/ProtobufUtils.h"
+#include "compute/ResultIterator.h"
+#include "memory/ArrowMemoryPool.h"
+#include "memory/ColumnarBatch.h"
+#include "operators/c2r/ArrowColumnarToRowConverter.h"
 #include "substrait/plan.pb.h"
-#include "utils/exception.h"
-#include "utils/metrics.h"
+#include "utils/Exception.h"
+#include "utils/Metrics.h"
 
 #ifdef GLUTEN_PRINT_DEBUG
 #include <iostream>
